Revert "feat: add RecentActivitiesCard component and integrate into Dashboard"
