+++ tsconfig.app.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/app.tsx","./src/env.d.ts","./src/main.tsx","./src/vite-env.d.ts","./src/components/deletereportmodal.tsx","./src/components/protectedroute.tsx","./src/components/respondreportmodal.tsx","./src/components/viewreportmodal.tsx","./src/components/layout/layout.tsx","./src/components/layout/sidebar.tsx","./src/components/report/createreportmodal.tsx","./src/components/report/deletereportmodal.tsx","./src/components/report/respondreportmodal.tsx","./src/components/report/viewreportmodal.tsx","./src/components/ui/accordion.tsx","./src/components/ui/alert-dialog.tsx","./src/components/ui/alert.tsx","./src/components/ui/aspect-ratio.tsx","./src/components/ui/avatar.tsx","./src/components/ui/badge.tsx","./src/components/ui/breadcrumb.tsx","./src/components/ui/button.tsx","./src/components/ui/calendar.tsx","./src/components/ui/card.tsx","./src/components/ui/carousel.tsx","./src/components/ui/chart.tsx","./src/components/ui/checkbox.tsx","./src/components/ui/collapsible.tsx","./src/components/ui/command.tsx","./src/components/ui/context-menu.tsx","./src/components/ui/dialog.tsx","./src/components/ui/drawer.tsx","./src/components/ui/dropdown-menu.tsx","./src/components/ui/form.tsx","./src/components/ui/hover-card.tsx","./src/components/ui/input-otp.tsx","./src/components/ui/input.tsx","./src/components/ui/label.tsx","./src/components/ui/menubar.tsx","./src/components/ui/navigation-menu.tsx","./src/components/ui/pagination.tsx","./src/components/ui/popover.tsx","./src/components/ui/progress.tsx","./src/components/ui/radio-group.tsx","./src/components/ui/resizable.tsx","./src/components/ui/scroll-area.tsx","./src/components/ui/select.tsx","./src/components/ui/separator.tsx","./src/components/ui/sheet.tsx","./src/components/ui/skeleton.tsx","./src/components/ui/slider.tsx","./src/components/ui/sonner.tsx","./src/components/ui/stats-card.tsx","./src/components/ui/switch.tsx","./src/components/ui/table.tsx","./src/components/ui/tabs.tsx","./src/components/ui/textarea.tsx","./src/components/ui/toast.tsx","./src/components/ui/toaster.tsx","./src/components/ui/toggle-group.tsx","./src/components/ui/toggle.tsx","./src/components/ui/tooltip.tsx","./src/hooks/use-toast.ts","./src/hooks/useauth.ts","./src/lib/api-routes.ts","./src/lib/api.ts","./src/lib/env.ts","./src/lib/utils.ts","./src/pages/agenda.tsx","./src/pages/banners.tsx","./src/pages/cashbacks.tsx","./src/pages/dashboard.tsx","./src/pages/login.tsx","./src/pages/microchips.tsx","./src/pages/payments.tsx","./src/pages/plans.tsx","./src/pages/register.tsx","./src/pages/reports.tsx","./src/pages/services.tsx","./src/pages/users.tsx","./src/services/banners-service.ts","./src/services/cashback-service.ts","./src/services/get-cashback-statistics.ts","./src/services/get-cashback-transactions.ts","./src/services/get-dashboard-overview.ts","./src/services/get-payment-stats.ts","./src/services/get-payments-list.ts","./src/services/get-recent-activity.ts","./src/services/get-scheduling-metrics.ts","./src/services/get-service-metrics.ts","./src/services/get-subscription-metrics.ts","./src/services/get-users.ts","./src/services/plans-service.ts","./src/services/reports-service.ts","./src/theme/colors.ts","./src/types/auth.ts","./src/types/banners.ts","./src/types/cashback.ts","./src/types/dashboard.ts","./src/types/payments.ts","./src/types/pets.ts","./src/types/plans.ts","./src/types/reports.ts","./src/types/subscriptions.ts","./src/types/users.ts"],"errors":true,"version":"5.8.3"}
+{"root":["./src/app.tsx","./src/env.d.ts","./src/main.tsx","./src/vite-env.d.ts","./src/components/protectedroute.tsx","./src/components/layout/layout.tsx","./src/components/layout/sidebar.tsx","./src/components/report/createreportmodal.tsx","./src/components/report/deletereportmodal.tsx","./src/components/report/respondreportmodal.tsx","./src/components/report/viewreportmodal.tsx","./src/components/ui/accordion.tsx","./src/components/ui/alert-dialog.tsx","./src/components/ui/alert.tsx","./src/components/ui/aspect-ratio.tsx","./src/components/ui/avatar.tsx","./src/components/ui/badge.tsx","./src/components/ui/breadcrumb.tsx","./src/components/ui/button.tsx","./src/components/ui/calendar.tsx","./src/components/ui/card.tsx","./src/components/ui/carousel.tsx","./src/components/ui/chart.tsx","./src/components/ui/checkbox.tsx","./src/components/ui/collapsible.tsx","./src/components/ui/command.tsx","./src/components/ui/context-menu.tsx","./src/components/ui/dialog.tsx","./src/components/ui/drawer.tsx","./src/components/ui/dropdown-menu.tsx","./src/components/ui/form.tsx","./src/components/ui/hover-card.tsx","./src/components/ui/input-otp.tsx","./src/components/ui/input.tsx","./src/components/ui/label.tsx","./src/components/ui/menubar.tsx","./src/components/ui/navigation-menu.tsx","./src/components/ui/pagination.tsx","./src/components/ui/popover.tsx","./src/components/ui/progress.tsx","./src/components/ui/radio-group.tsx","./src/components/ui/resizable.tsx","./src/components/ui/scroll-area.tsx","./src/components/ui/select.tsx","./src/components/ui/separator.tsx","./src/components/ui/sheet.tsx","./src/components/ui/skeleton.tsx","./src/components/ui/slider.tsx","./src/components/ui/sonner.tsx","./src/components/ui/stats-card.tsx","./src/components/ui/switch.tsx","./src/components/ui/table.tsx","./src/components/ui/tabs.tsx","./src/components/ui/textarea.tsx","./src/components/ui/toast.tsx","./src/components/ui/toaster.tsx","./src/components/ui/toggle-group.tsx","./src/components/ui/toggle.tsx","./src/components/ui/tooltip.tsx","./src/hooks/use-toast.ts","./src/hooks/useauth.ts","./src/lib/api-routes.ts","./src/lib/api.ts","./src/lib/env.ts","./src/lib/utils.ts","./src/pages/agenda.tsx","./src/pages/banners.tsx","./src/pages/cashbacks.tsx","./src/pages/dashboard.tsx","./src/pages/login.tsx","./src/pages/microchips.tsx","./src/pages/payments.tsx","./src/pages/plans.tsx","./src/pages/register.tsx","./src/pages/reports.tsx","./src/pages/services.tsx","./src/pages/users.tsx","./src/services/banners-service.ts","./src/services/get-dashboard-overview.ts","./src/services/get-payment-stats.ts","./src/services/get-payments-list.ts","./src/services/get-recent-activity.ts","./src/services/get-scheduling-metrics.ts","./src/services/get-service-metrics.ts","./src/services/get-subscription-metrics.ts","./src/services/get-users.ts","./src/services/reports-service.ts","./src/theme/colors.ts","./src/types/auth.ts","./src/types/banners.ts","./src/types/dashboard.ts","./src/types/payments.ts","./src/types/pets.ts","./src/types/reports.ts","./src/types/subscriptions.ts","./src/types/users.ts"],"errors":true,"version":"5.8.3"}
